sec-llm-agents/requirements.txt
crewai==0.51.1
langchain-openai==0.2.1
langchain-core==0.3.8
pydantic==2.9.2
fastapi==0.115.0
uvicorn[standard]==0.30.6
gitpython==3.1.43
requests==2.32.3
python-multipart==0.0.9
typer==0.12.5
rich==13.9.3
